requirements.txt
pynest-api
ccxt
loguru
uvicorn
python-dotenv
pydantic
sqlalchemy
asyncpg
